content/canvas/test/test_2d.line.width.invalid.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.line.width.invalid</title>
<!-- Testing: Setting lineWidth to invalid values is ignored -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;
try {

ctx.lineWidth = 1.5;
ok(ctx.lineWidth === 1.5, "ctx.lineWidth === 1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = 0;
todo(ctx.lineWidth === 1.5, "ctx.lineWidth === 1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = -1;
todo(ctx.lineWidth === 1.5, "ctx.lineWidth === 1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = Infinity;
todo(ctx.lineWidth === 1.5, "ctx.lineWidth === 1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = -Infinity;
todo(ctx.lineWidth === 1.5, "ctx.lineWidth === 1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = NaN;
todo(ctx.lineWidth === 1.5, "ctx.lineWidth === 1.5");

} catch (e) {
    _thrown_outer = true;
}
todo(!_thrown_outer, 'should not throw exception');

SimpleTest.finish();

});
</script>
